requirements.txt
PyYAML==3.10
alphasign==1.0
google-api-python-client==1.0
httplib2==0.7.7
pyserial==2.6
python-dateutil==2.1
python-gflags==2.0
pytz==2012j
pyusb==1.0.0a3
six==1.2.0
wsgiref==0.1.2
